Allow importing Perses dashboads in the UI
This splits the MigrateView into two separate flows, one for Grafana and
one for Perses. The Grafana flow is unchanged, but the Perses flow
skips the migrate translation step, going straight to the
project select and import.
